Clean update of front according to the sonar plugin exemple method
