+++ conf/webpack/webpack.config.prod.js
@@ -1,0 +1,69 @@
+/*
+ * Copyright (C) 2009-2020 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+const webpack = require('webpack');
+const config = require('./webpack.config');
+const getClientEnvironment = require('../env');
+
+// Get environment variables to inject into our app.
+const env = getClientEnvironment();
+
+// Assert this just to be safe.
+// Development builds of React are slow and not intended for production.
+if (env['process.env.NODE_ENV'] !== '"production"') {
+  throw new Error('Production builds must have NODE_ENV=production.');
+}
+
+const noUglify = process.argv.some(arg => arg.indexOf('--no-uglify') > -1);
+
+// Don't attempt to continue if there are any errors.
+config.bail = true;
+
+config.plugins = [
+  // Makes some environment variables available to the JS code, for example:
+  // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
+  // It is absolutely essential that NODE_ENV was set to production here.
+  // Otherwise React will be compiled in the very slow development mode.
+  new webpack.DefinePlugin(env),
+
+  // This helps ensure the builds are consistent if source hasn't changed:
+  new webpack.optimize.OccurrenceOrderPlugin(),
+
+  // Try to dedupe duplicated modules, if any:
+  new webpack.optimize.DedupePlugin()
+];
+
+if (!noUglify) {
+  config.plugins.push(
+    new webpack.optimize.UglifyJsPlugin({
+      compress: {
+        screw_ie8: true, // React doesn't support IE8
+        warnings: false
+      },
+      mangle: {
+        screw_ie8: true
+      },
+      output: {
+        comments: false,
+        screw_ie8: true
+      }
+    })
+  );
+}
+
+module.exports = config;
