+++ package.json
@@ -1,45 +1,56 @@
 {
-    "name": "sonar-cnes-report-plugin",
+    "name": "sonar-cnes-report",
     "license": "GPL-3.0",
-    "version": "4.2.0",
+    "version": "5.0.0",
     "devDependencies": {
-        "@babel/core": "^7.8.7",
-        "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/preset-env": "^7.8.7",
-        "@babel/preset-react": "^7.8.3",
-        "@emotion/core": "^10.0.17",
-        "@emotion/styled": "^10.0.17",
-        "autoprefixer": "9.7.4",
+        "autoprefixer": "6.2.2",
         "babel-core": "6.14.0",
-        "babel-jest": "^25.1.0",
-        "babel-loader": "^8.0.6",
+        "babel-loader": "6.2.5",
         "babel-preset-react-app": "0.2.1",
-        "css-loader": "^3.4.2",
-        "emotion-theming": "^10.0.19",
-        "lodash": "4.17.21",
-        "postcss-calc": "7.0.2",
-        "postcss-custom-properties": "9.1.1",
-        "postcss-loader": "3.0.0",
-        "react": "16.13.0",
-        "react-dev-utils": "^11.0.4",
-        "react-dom": "16.13.0",
-        "react-router": "3.2.6",
-        "sonar-ui-common": "1.0.33",
-        "style-loader": "1.1.3",
-        "uglifyjs-webpack-plugin": "^2.2.0",
-        "webpack": "^4.42.0"
+        "cross-env": "2.0.0",
+        "cross-spawn": "4.0.0",
+        "css-loader": "0.23.1",
+        "detect-port": "1.0.0",
+        "dotenv": "2.0.0",
+        "enzyme": "2.6.0",
+        "enzyme-to-json": "1.4.5",
+        "expose-loader": "0.7.1",
+        "express": "4.13.4",
+        "express-http-proxy": "0.6.0",
+        "filesize": "3.3.0",
+        "find-cache-dir": "0.1.1",
+        "gzip-size": "3.0.0",
+        "imports-loader": "0.6.5",
+        "json-loader": "0.5.4",
+        "path-exists": "2.1.0",
+        "postcss-loader": "0.8.0",
+        "prettier": "0.22.0",
+        "react": "15.6.2",
+        "react-addons-shallow-compare": "15.6.2",
+        "react-addons-test-utils": "15.6.2",
+        "react-dev-utils": "0.2.1",
+        "react-dom": "15.6.2",
+        "react-router": "3.0.2",
+        "react-transform-hmr": "1.0.4",
+        "recursive-readdir": "2.1.0",
+        "rimraf": "2.5.4",
+        "script-loader": "0.6.1",
+        "strip-ansi": "3.0.1",
+        "style-loader": "0.13.0",
+        "webpack": "1.13.2",
+        "webpack-dev-server": "1.16.1"
     },
     "scripts": {
-        "build": "node src/main/js/scripts/build.js"
+        "build": "node scripts/build.js"
     },
     "babel": {
         "presets": [
-            "@babel/env",
-            "@babel/react"
-        ],
-        "plugins": [
-            "@babel/plugin-proposal-class-properties"
+            "react-app"
         ]
     },
-    "dependencies": {}
+    "dependencies": {
+        "backbone": "^1.4.0",
+        "jquery": "^1.11.0",
+        "underscore": "^1.9.1"
+    }
 }
