--- src/main/js/style.css
@@ -1,22 +1,0 @@
-.forminput select, .forminput input{
-    width:100%;
-}
-
-.login-label{
-    display: block;
-}
-
-.info-message{
-    color:grey;
-}
-
-.compatibility-warning {
-    padding-bottom: inherit;
-    color: red;
-    font-style: italic;
-}
-
-.page-simple {
-  width: 30%;
-  margin: auto;
-}
